requirements.txt
# Production requirements for Attention Pulse Pipeline
# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=8.3.0
deeplake>=3.10.0


# Computer vision and augmentation
albumentations>=1.3.0
timm>=0.6.0

# Hyperparameter optimization and experiment tracking
optuna>=3.0.0
mlflow>=2.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pathlib>=1.0.0
typing-extensions>=4.0.0

fastapi>=0.95.0    
uvicorn>=0.18.0    
python-multipart>=0.0.5  

onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0  

requests>=2.28.0
python-dotenv>=1.0.0

torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12  # For MobileViT and EfficientNet
onnx>=1.15.0
onnxruntime>=1.16.3

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0
mediapipe>=0.10.8
# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

gdown>=4.7.1
kaggle>=1.5.16

focal-loss-torch>=0.0.1

tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0

jupyter>=1.0.0
ipywidgets>=8.0.0

